Spliter les components de confirmation
