+++ client/src/components/steps/components/Confirmation/ClientConfirmationStep.tsx
@@ -1,0 +1,5 @@
+import { BaseConfirmationStep, BaseConfirmationStepProps } from "./BaseConfirmationStep";
+
+export function ClientConfirmationStep({ form }: Readonly<BaseConfirmationStepProps>) {
+    return <BaseConfirmationStep form={form} />;
+  }
