+++ client/src/components/steps/components/Confirmation/SupplierConfirmationStep.tsx
@@ -1,0 +1,5 @@
+import { BaseConfirmationStep, BaseConfirmationStepProps } from "./BaseConfirmationStep";
+
+export function SupplierConfirmationStep({ form }: Readonly<BaseConfirmationStepProps>) {
+    return <BaseConfirmationStep form={form} />;
+  }
