+++ client/src/components/steps/components/Delivery/DeliveryDetailsStep.tsx
@@ -1,10 +1,10 @@
 import { UseFormReturn } from "react-hook-form";
 import { Input } from "@/components/ui/input";
 import { motion } from "framer-motion";
-import { AuthForm } from "../../../pages/types";
+import { AuthForm } from "../../../../pages/types";
 import { useTranslation } from 'react-i18next';
 import { CustomFileUpload } from "./CustomFileUpload";
-import { DocumentProgress } from "../DetailsStep";
+import { DocumentProgress } from "../../DetailsStep";
 
 interface DeliveryDetailsStepProps {
   readonly form: UseFormReturn<AuthForm>;
