+++ client/src/locales/en/auth.json
@@ -11,7 +11,8 @@
     "username": "Username",
     "password": "Password",
     "confirmPassword": "Confirm password",
-    "role": "Role"
+    "role": "Role",
+    "address":"Address"
   },
   "placeholders": {
     "firstName": "Enter your first name",
