+++ client/src/locales/fr/auth.json
@@ -11,7 +11,8 @@
     "username": "Nom d'utilisateur",
     "password": "Mot de passe",
     "confirmPassword": "Confirmer le mot de passe",
-    "role": "Rôle"
+    "role": "Rôle",
+    "address":"Adresse"
   },
   "placeholders": {
     "firstName": "Entrez votre prénom",
